Move TextSample to experimental
